pyproject.toml
[project]
name = "agent-template"
version = "0.1.0"
description = "AIOps Agents template built on Google ADK"
authors = [
  {name = "AIOps Team", email = "aiops-team@g.globo"}
]
requires-python = ">=3.13"
dependencies = [
  "google-adk[eval]>=1.18.0",
  "uv>=0.8.4",
  "immutabledict>=4.2.1,<5",
  "regex>=2024.11.6,<2025",
  "sqlglot>=26.10.1,<27",
  "db-dtypes>=1.4.2,<2",
  "absl-py>=2.2.2,<3",
  "litellm>=1.80.9",
  "sentence-transformers>=5.1.2",
  "rank-bm25>=0.2.2",
  "langchain-text-splitters>=1.0.0",
  "chromadb>=1.3.6",
]

[tool.setuptools]
packages = ["app"]
